Moves Terminal Converter to Expressions Language
The Expressions Language defines the Hex Terminal Literal, thus the
corresponding terminal converter shouldbe part of the expressions
language aswell
