+++ plugins/org.yakindu.sct.model.stext/META-INF/MANIFEST.MF
@@ -25,7 +25,7 @@ Import-Package: org.apache.commons.logging,
  org.eclipse.xtext.xbase.lib
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Export-Package: org.yakindu.sct.model.stext,
- org.yakindu.sct.model.stext.conversion,
+ org.yakindu.sct.model.stext.expressions,
  org.yakindu.sct.model.stext.formatting,
  org.yakindu.sct.model.stext.inferrer,
  org.yakindu.sct.model.stext.naming,
@@ -39,7 +39,6 @@ Export-Package: org.yakindu.sct.model.stext,
  org.yakindu.sct.model.stext.stext,
  org.yakindu.sct.model.stext.stext.impl,
  org.yakindu.sct.model.stext.stext.util,
- org.yakindu.sct.model.stext.expressions,
  org.yakindu.sct.model.stext.util,
  org.yakindu.sct.model.stext.validation
 Bundle-ClassPath: .
